mptcp: Fix Bug #41: Handle missing entries in the fullmesh-path-manager
It may be that the address of the initial subflow is not yet part of the
fullmesh path-manager's list of addresses. May happen because the
path-manager waits 500ms before adding the address.

In general, we do not handle a missing address in the path-manager's
list not correctly. We always assume that it is there.

Change everything, so that get_local_id returns -1 if the address is not
part of the fullmesh list. In case -1 is returned on a MP_JOIN, we
simply ignore the SYN.

In case -1 is returned on a MP_CAPABLE, we disable the path-manager and
fall back to the 'default'-path-management, which basically does
nothing.

Fixes: Bug #41 (https://github.com/multipath-tcp/mptcp/issues/41)
Fixes: c31d21d039d80 (Modularize the path-manager)

+++ net/mptcp/mptcp_ndiffports.c
@@ -82,7 +82,7 @@ exit:
 	sock_put(meta_sk);
 }
 
-static void ndiffports_new_session(struct sock *meta_sk, u8 id)
+static void ndiffports_new_session(struct sock *meta_sk, int id)
 {
 	struct mptcp_cb *mpcb = tcp_sk(meta_sk)->mpcb;
 	struct ndiffports_priv *fmp = (struct ndiffports_priv *)&mpcb->mptcp_pm[0];
